src/GrabYourTime/src/GrabYourTime/requirements.txt
amqp==2.2.2
django-crontab==0.7.1
billiard==3.5.0.3
celery==4.2.1
certifi==2017.11.5
Django==2.1.1
django-markdown-deux==1.0.5
django-model-utils==3.0.0
django-notifications-hq==1.4
django-pagedown==1.0.6
jsonfield==2.0.2
kombu==4.2.1
markdown2==2.3.5
pytz==2017.3
sqlparse==0.2.4
vine==1.1.4
django-phone-field==1.6.0
twilio==6.20.0
